Refactoring after suggestions for best practices and performance improvements
