+++ components/Button.tsx
@@ -1,30 +1,34 @@
-// components/Hello.tsx
 import React from 'react';
-import {StyleSheet, Text, TouchableOpacity} from 'react-native';
-import {Colors} from 'react-native/Libraries/NewAppScreen';
+import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+import { Colors } from 'react-native/Libraries/NewAppScreen';
 
 export interface ButtonProps {
     label: String
 }
 
-export class Button extends React.Component<ButtonProps> {
-    constructor(props: ButtonProps) {
-        super(props);
-    }
-
+export class Button extends React.PureComponent<ButtonProps> {
     render() {
         return (
-            <TouchableOpacity style={styles.button}>
-                <Text style={styles.text}>{this.props.label}</Text>
-            </TouchableOpacity>
+            <View style={styles.wrapper}>
+                <TouchableOpacity style={styles.button}>
+                    <Text style={styles.buttonText}>{this.props.label}</Text>
+                </TouchableOpacity>
+            </View>
         );
     }
 }
 
-// styles
 const styles = StyleSheet.create({
+    wrapper: {
+        width: '100%',
+        alignItems: 'center',
+        alignSelf: 'center',
+        backgroundColor: Colors.white,
+        paddingHorizontal: 30,
+        paddingVertical: 20
+    },
     button: {
-        width: "100%",
+        width: '100%',
         alignItems: 'center',
         alignSelf: 'center',
         backgroundColor: 'blue',
@@ -34,7 +38,7 @@ const styles = StyleSheet.create({
         paddingVertical: 10,
         margin: 30
     },
-    text: {
+    buttonText: {
         color: Colors.white,
         fontWeight: 'bold',
         fontSize: 20,
